requirements.txt
aiohttp~=3.9.1
inflection~=0.5.1
pyjwt~=2.8.0
python-dateutil~=2.8.1
python-smarttub~=0.0.37
influxdb
pyyaml
requests
